# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GigShield - Gig Worker Insurance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workers/{id}:
    get:
      operationId: getWorker
      tags: [workers]
      summary: Get a worker by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Worker found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "404":
          description: Worker not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateWorker
      tags: [workers]
      summary: Update a worker profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkerRequest"
      responses:
        "200":
          description: Worker updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
  /workers/{id}/dashboard:
    get:
      operationId: getWorkerDashboard
      tags: [workers]
      summary: Get worker dashboard data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardData"
  /policies:
    get:
      operationId: listPolicies
      tags: [policies]
      summary: List available insurance plans
      responses:
        "200":
          description: List of available plans
          content:
            application/json:
              schema:
                type: object
                properties:
                  plans:
                    type: array
                    items:
                      $ref: "#/components/schemas/Plan"
                required:
                  - plans
    post:
      operationId: createPolicy
      tags: [policies]
      summary: Subscribe to an insurance plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyRequest"
      responses:
        "201":
          description: Policy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /policies/{id}:
    get:
      operationId: getPolicy
      tags: [policies]
      summary: Get a specific policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Policy found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
    delete:
      operationId: cancelPolicy
      tags: [policies]
      summary: Cancel a policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Policy cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
  /workers/{workerId}/policies:
    get:
      operationId: getWorkerPolicies
      tags: [policies]
      summary: Get all policies for a worker
      parameters:
        - name: workerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Worker policies
          content:
            application/json:
              schema:
                type: object
                properties:
                  policies:
                    type: array
                    items:
                      $ref: "#/components/schemas/Policy"
                required:
                  - policies
...
  /triggers:
    get:
      operationId: listTriggers
      tags: [triggers]
      summary: List recent trigger events
      parameters:
        - name: zone
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of triggers
          content:
            application/json:
              schema:
                type: object
                properties:
                  triggers:
                    type: array
                    items:
                      $ref: "#/components/schemas/TriggerEvent"
                required:
                  - triggers
    post:
      operationId: createTrigger
      tags: [triggers]
      summary: Create a trigger event (system/admin use)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTriggerRequest"
      responses:
        "201":
          description: Trigger created and claims processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerResult"
...
  /admin/triggers:
    get:
      operationId: adminListTriggers
      tags: [admin]
      summary: Admin - list all triggers
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Triggers list
          content:
            application/json:
              schema:
                type: object
                properties:
                  triggers:
                    type: array
                    items:
                      $ref: "#/components/schemas/TriggerEvent"
                  total:
                    type: integer
                required:
                  - triggers
                  - total
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    Worker:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        zone:
          type: string
        platform:
          type: string
        trustScore:
          type: number
          format: float
        riskScore:
          type: number
          format: float
        isActive:
          type: boolean
        role:
          type: string
          enum: [worker, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - zone
        - platform
        - trustScore
        - riskScore
        - isActive
        - role
        - createdAt
...
    UpdateWorkerRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        zone:
          type: string
        platform:
          type: string
    Plan:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        weeklyPremium:
          type: number
          format: float
        maxPayoutPerWeek:
          type: number
          format: float
        coverageTypes:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - description
        - weeklyPremium
        - maxPayoutPerWeek
        - coverageTypes
        - features
    Policy:
      type: object
      properties:
        id:
          type: integer
        workerId:
          type: integer
        planId:
          type: string
        planName:
          type: string
        status:
          type: string
          enum: [active, expired, cancelled]
        weeklyPremium:
          type: number
          format: float
        maxPayoutPerWeek:
          type: number
          format: float
        zone:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - workerId
        - planId
        - planName
        - status
        - weeklyPremium
        - maxPayoutPerWeek
        - zone
        - startDate
        - createdAt
    CreatePolicyRequest:
      type: object
      properties:
        workerId:
          type: integer
        planId:
          type: string
        zone:
          type: string
      required:
        - workerId
        - planId
        - zone
    Claim:
      type: object
      properties:
        id:
          type: integer
        workerId:
          type: integer
        policyId:
          type: integer
        triggerId:
          type: integer
        amount:
          type: number
          format: float
        status:
          type: string
          enum: [pending, approved, rejected, paid]
        triggerType:
          type: string
        triggerDescription:
          type: string
        zone:
          type: string
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - workerId
        - policyId
        - triggerId
        - amount
        - status
        - triggerType
        - triggerDescription
        - zone
        - createdAt
...
    TriggerEvent:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [rain, heat, curfew, platform_outage, flood, storm]
        zone:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        rainfallMm:
          type: number
          format: float
          nullable: true
        temperatureCelsius:
          type: number
          format: float
          nullable: true
        orderDropPercent:
          type: number
          format: float
          nullable: true
        affectedWorkers:
          type: integer
        totalPayout:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - zone
        - description
        - severity
        - affectedWorkers
        - totalPayout
        - createdAt
    CreateTriggerRequest:
      type: object
      properties:
        type:
          type: string
          enum: [rain, heat, curfew, platform_outage, flood, storm]
        zone:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        rainfallMm:
          type: number
          format: float
        temperatureCelsius:
          type: number
          format: float
        orderDropPercent:
          type: number
          format: float
      required:
        - type
        - zone
        - description
        - severity
    TriggerResult:
      type: object
      properties:
        trigger:
          $ref: "#/components/schemas/TriggerEvent"
        claimsCreated:
          type: integer
        totalPayout:
          type: number
          format: float
      required:
        - trigger
        - claimsCreated
        - totalPayout
...
    DashboardData:
      type: object
      properties:
        worker:
          $ref: "#/components/schemas/Worker"
        activePolicy:
          $ref: "#/components/schemas/Policy"
          nullable: true
        recentClaims:
          type: array
          items:
            $ref: "#/components/schemas/Claim"
        totalEarningsProtected:
          type: number
          format: float
        totalClaims:
          type: integer
        totalPaid:
          type: number
          format: float
        recentTriggers:
          type: array
          items:
            $ref: "#/components/schemas/TriggerEvent"
        coverageStatus:
          type: string
          enum: [active, inactive, pending]
      required:
        - worker
        - recentClaims
        - totalEarningsProtected
        - totalClaims
        - totalPaid
        - recentTriggers
        - coverageStatus
...