# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GigShield - Gig Worker Insurance API
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: string;
  name: string;
  description: string;
  weeklyPremium: number;
  maxPayoutPerWeek: number;
  coverageTypes: string[];
  features: string[];
}